add gates

+++ app/Http/Controllers/ProductController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\ProductCreateRequest;
 use App\Http\Resources\ProductResource;
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Symfony\Component\HttpFoundation\Response;
 
 class ProductController extends Controller
@@ -13,6 +14,8 @@ class ProductController extends Controller
 
     public function index()
     {
+        Gate::authorize('view', 'products');
+
         $products = Product::paginate();
 
         return ProductResource::collection($products);
@@ -21,6 +24,8 @@ class ProductController extends Controller
 
     public function store(ProductCreateRequest $request)
     {
+        Gate::authorize('edit', 'products');
+
         $product = Product::create($request->only('title', 'description', 'image', 'price'));
 
         return response($product, Response::HTTP_CREATED);
@@ -29,12 +34,16 @@ class ProductController extends Controller
 
     public function show($id)
     {
+        Gate::authorize('view', 'products');
+
         return new ProductResource(Product::find($id));
     }
 
 
     public function update(Request $request, $id)
     {
+        Gate::authorize('edit', 'products');
+
         $product = Product::find($id);
 
         $product->update($request->only('title', 'description', 'image', 'price'));
@@ -45,6 +54,8 @@ class ProductController extends Controller
 
     public function destroy($id)
     {
+        Gate::authorize('edit', 'products');
+
         Product::destroy($id);
 
         return response(null, Response::HTTP_NO_CONTENT);
